requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
multidict==6.0.4
pya3rt==1.1
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
gunicorn